# Improve performance of SourceText.IsBinary on .NET Core