@jaredpar 2022-05-05
Can you share out the benchmark you wrote here? Mostly for my own education.