@Neme12 2022-05-05
@jaredpar

This is the second benchmark:
```xml
<Project Sdk="Microsoft.NET.Sdk">

  <PropertyGroup>
    <OutputType>Exe</OutputType>
    <TargetFrameworks>net6.0;netcoreapp3.1;net472</TargetFrameworks>
    <ImplicitUsings>enable</ImplicitUsings>
    <Nullable>enable</Nullable>
  </PropertyGroup>

  <ItemGroup>
    <PackageReference Include="BenchmarkDotNet" Version="0.13.1" />
  </ItemGroup>

  <ItemGroup Condition="'$(TargetFramework)' != 'net472'">
    <PackageReference Include="System.Collections" Version="4.3.0" />
    <PackageReference Include="System.IO.FileSystem.Primitives" Version="4.3.0" />
    <PackageReference Include="System.Runtime.Extensions" Version="4.3.0" />
    <PackageReference Include="System.Runtime.Handles" Version="4.3.0" />
    <PackageReference Include="System.Runtime.InteropServices" Version="4.3.0" />
  </ItemGroup>

  <ItemGroup>
    <ProjectReference Include="..\src\Compilers\CSharp\Portable\Microsoft.CodeAnalysis.CSharp.csproj" />
  </ItemGroup>

</Project>
```
```c#
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Jobs;
using BenchmarkDotNet.Running;
using Microsoft.CodeAnalysis.Text;

BenchmarkRunner.Run<SourceTextBenchmark>();

[SimpleJob(RuntimeMoniker.Net472)]
[SimpleJob(RuntimeMoniker.NetCoreApp31)]
[SimpleJob(RuntimeMoniker.Net60)]
public class SourceTextBenchmark
{
    private readonly string _largerTextPath =
        @"C:\Users\{username}\source\repos\dotnet\roslyn\src\Compilers\CSharp\Portable\Generated\BoundNodes.xml.Generated.cs"; // 902 KB;

    private FileStream? _fileStream;

    [IterationSetup]
    public void Setup()
    {
        _fileStream = File.OpenRead(_largerTextPath);
    }

    [IterationCleanup]
    public void Cleanup()
    {
        _fileStream!.Dispose();
        _fileStream = null;
    }

    [Benchmark]
    public SourceText SourceTextFrom()
    {
        return SourceText.From(_fileStream!, throwIfBinaryDetected: true);
    }
}
```

I don't have the source for the first benchmark. First, I coped the `IsBinary` method to check the new implementation and the benchmark just had 2 methods, one called the original `IsBinary` method and the other one called `IsBinary2`.